chore(parser): move getOperationResponseHeader to common parser
