+++ src/openApi/v3/parser/operation.ts
@@ -1,7 +1,7 @@
 import type { Operation, OperationParameter, OperationParameters } from '../../../types/client';
 import type { Config } from '../../../types/config';
 import { getRef } from '../../common/parser/getRef';
-import { getOperationName } from '../../common/parser/operation';
+import { getOperationName, getOperationResponseHeader } from '../../common/parser/operation';
 import { toSortedByRequired } from '../../common/parser/sort';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiOperation } from '../interfaces/OpenApiOperation';
@@ -9,7 +9,6 @@ import type { OpenApiRequestBody } from '../interfaces/OpenApiRequestBody';
 import { getOperationErrors } from './getOperationErrors';
 import { getOperationParameters } from './getOperationParameters';
 import { getOperationRequestBody } from './getOperationRequestBody';
-import { getOperationResponseHeader } from './getOperationResponseHeader';
 import { getOperationResponses } from './getOperationResponses';
 import { getOperationResults } from './getOperationResults';
 import { getServiceName } from './service';
